Crate egglog_reports

Structs

IterationReport

Plan

RuleReport

RuleSetReport

RunReport

Running a schedule produces a report of the results. This includes rough timing information and whether the database was updated. Calling union on two run reports adds the timing information together.

Scan

SingleScan

StageStats

Enums

ReportLevel

Stage